+++ apps/terriamap/lib/Views/terriaStore.ts
@@ -19,10 +19,9 @@ class TerriaStore {
   }
 
   async init() {
-    //@ts-expect-error: need to convert to TS
     await import("terriajs/lib/Core/prerequisites");
 
-    const { terria, viewState } = await import("../../index.js").then(
+    const { terria, viewState } = await import("../../index").then(
       (module) => module.default
     );