+++ packages/terriajs/lib/ReactViews/SidePanel/SidePanel.tsx
@@ -177,7 +177,7 @@ const SidePanel = observer<React.FC<SidePanelProps>>(
         >
           <SearchBoxAndResults
             placeholder={applyTranslationIfExists(
-              terria.searchBarModel.placeholder,
+              terria.searchBarModel.config.placeholder,
               i18n
             )}
           />