+++ packages/terriajs/test/Models/CesiumSpec.ts
@@ -31,6 +31,7 @@ import createStratumInstance from "../../lib/Models/Definition/createStratumInst
 import updateModelFromJson from "../../lib/Models/Definition/updateModelFromJson";
 import upsertModelFromJson from "../../lib/Models/Definition/upsertModelFromJson";
 import Terria from "../../lib/Models/Terria";
+import { ConfigStrata } from "../../lib/Models/Config/ConfigStrata";
 import { RectangleTraits } from "../../lib/Traits/TraitsClasses/MappableTraits";
 import TerriaViewer from "../../lib/ViewModels/TerriaViewer";
 import { worker } from "../mocks/browser";
@@ -348,10 +349,11 @@ describeIfSupported("Cesium Model", function () {
         cesium = new Cesium(terria.mainViewer, container);
         scene = cesium.scene;
         cesium.terriaViewer.viewerOptions.useTerrain = true;
-        terria.configParameters.cesiumTerrainAssetId = 123;
-        terria.configParameters.cesiumTerrainUrl =
-          "https://cesium-terrain.example.com/";
-        terria.configParameters.useCesiumIonTerrain = true;
+        terria.configParameters.update(ConfigStrata.definition, {
+          cesiumTerrainAssetId: 123,
+          cesiumTerrainUrl: "https://cesium-terrain.example.com/",
+          useCesiumIonTerrain: true
+        });
 
         workbenchTerrainItem = upsertModelFromJson(
           CatalogMemberFactory,
@@ -422,7 +424,11 @@ describeIfSupported("Cesium Model", function () {
       runInAction(() => {
         cesium.terriaViewer.viewerOptions.useTerrain = true;
         terria.workbench.removeAll();
-        terria.configParameters.cesiumTerrainAssetId = undefined;
+        terria.configParameters.setValue(
+          ConfigStrata.definition,
+          "cesiumTerrainAssetId",
+          undefined
+        );
       });
 
       await terrainLoadPromise(cesium);
@@ -439,8 +445,16 @@ describeIfSupported("Cesium Model", function () {
       runInAction(() => {
         cesium.terriaViewer.viewerOptions.useTerrain = true;
         terria.workbench.removeAll();
-        terria.configParameters.cesiumTerrainAssetId = undefined;
-        terria.configParameters.cesiumTerrainUrl = undefined;
+        terria.configParameters.setValue(
+          ConfigStrata.definition,
+          "cesiumTerrainAssetId",
+          undefined
+        );
+        terria.configParameters.setValue(
+          ConfigStrata.definition,
+          "cesiumTerrainUrl",
+          undefined
+        );
       });
 
       await terrainLoadPromise(cesium);
@@ -453,9 +467,21 @@ describeIfSupported("Cesium Model", function () {
       runInAction(() => {
         cesium.terriaViewer.viewerOptions.useTerrain = true;
         terria.workbench.removeAll();
-        terria.configParameters.cesiumTerrainAssetId = undefined;
-        terria.configParameters.cesiumTerrainUrl = undefined;
-        terria.configParameters.useCesiumIonTerrain = false;
+        terria.configParameters.setValue(
+          ConfigStrata.definition,
+          "cesiumTerrainAssetId",
+          undefined
+        );
+        terria.configParameters.setValue(
+          ConfigStrata.definition,
+          "cesiumTerrainUrl",
+          undefined
+        );
+        terria.configParameters.setValue(
+          ConfigStrata.definition,
+          "useCesiumIonTerrain",
+          false
+        );
       });
 
       await terrainLoadPromise(cesium);
@@ -493,7 +519,11 @@ describeIfSupported("Cesium Model", function () {
     it("should throw a warning when cesiumIonAccessToken is invalid", async function () {
       runInAction(() => {
         // Set an invalid token for the test
-        terria2.configParameters.cesiumIonAccessToken = "expired_token";
+        terria2.configParameters.setValue(
+          ConfigStrata.definition,
+          "cesiumIonAccessToken",
+          "expired_token"
+        );
       });
       // Instantiate Cesium object with the invalid token
       cesium2 = new Cesium(terriaViewer2, container2);
@@ -515,7 +545,11 @@ describeIfSupported("Cesium Model", function () {
       expect(terriaViewer2.viewerOptions.useTerrain).toBe(true, "1");
       runInAction(() => {
         // Set an invalid token for the test
-        terria2.configParameters.cesiumIonAccessToken = "expired_token";
+        terria2.configParameters.setValue(
+          ConfigStrata.definition,
+          "cesiumIonAccessToken",
+          "expired_token"
+        );
       });
 
       // Instantiate Cesium object with the invalid token
@@ -532,9 +566,17 @@ describeIfSupported("Cesium Model", function () {
     it("should throw a warning when `cesiumIonAccessToken` is invalid and `cesiumTerrainAssetId` is present", async function () {
       runInAction(() => {
         // Set an invalid token for the test
-        terria2.configParameters.cesiumIonAccessToken = "expired_token";
+        terria2.configParameters.setValue(
+          ConfigStrata.definition,
+          "cesiumIonAccessToken",
+          "expired_token"
+        );
         // Set a valid asset id
-        terria2.configParameters.cesiumTerrainAssetId = 480278;
+        terria2.configParameters.setValue(
+          ConfigStrata.definition,
+          "cesiumTerrainAssetId",
+          480278
+        );
       });
       // Instantiate Cesium object with the invalid token and valid asset id
       cesium2 = new Cesium(terriaViewer2, container2);
@@ -561,8 +603,11 @@ describeIfSupported("Cesium Model", function () {
       );
 
       runInAction(() => {
-        terria2.configParameters.cesiumTerrainUrl =
-          "https://storage.googleapis.com/vic-datasets-public/xxxxxxx-xxxx-xxxx-xxxx-xxxxxxx/v1"; // An invalid url
+        terria2.configParameters.setValue(
+          ConfigStrata.definition,
+          "cesiumTerrainUrl",
+          "https://storage.googleapis.com/vic-datasets-public/xxxxxxx-xxxx-xxxx-xxxx-xxxxxxx/v1" // An invalid url
+        );
       });
       // Instantiate Cesium object with the invalid terrain url
       cesium2 = new Cesium(terriaViewer2, container2);