+++ packages/terriajs/test/Models/SearchProviders/LocationSearchProviderSpec.ts
@@ -1,7 +1,9 @@
+import { runInAction } from "mobx";
 import LocationSearchProviderMixin from "../../../lib/ModelMixins/SearchProviders/LocationSearchProviderMixin";
 import CommonStrata from "../../../lib/Models/Definition/CommonStrata";
 import BingMapsSearchProvider from "../../../lib/Models/SearchProviders/BingMapsSearchProvider";
 import Terria from "../../../lib/Models/Terria";
+import { ConfigStrata } from "../../../lib/Models/Config/TerriaConfig";
 
 describe("LocationSearchProvider", function () {
   let terria: Terria;
@@ -33,29 +35,41 @@ describe("LocationSearchProvider", function () {
 
   describe("searchBarModel fallback", function () {
     it(" - reads recommendedListLength from searchBarModel when not set on provider", function () {
-      terria.searchBarModel.setTrait(
-        CommonStrata.definition,
-        "recommendedListLength",
-        10
-      );
+      runInAction(() => {
+        terria.configParameters.mergeValue(
+          ConfigStrata.user,
+          "searchBarConfig",
+          {
+            recommendedListLength: 10
+          }
+        );
+      });
       expect(bingMapsSearchProvider.recommendedListLength).toEqual(10);
     });
 
     it(" - reads flightDurationSeconds from searchBarModel when not set on provider", function () {
-      terria.searchBarModel.setTrait(
-        CommonStrata.definition,
-        "flightDurationSeconds",
-        3.0
-      );
+      runInAction(() => {
+        terria.configParameters.mergeValue(
+          ConfigStrata.user,
+          "searchBarConfig",
+          {
+            flightDurationSeconds: 3.0
+          }
+        );
+      });
       expect(bingMapsSearchProvider.flightDurationSeconds).toEqual(3.0);
     });
 
     it(" - provider recommendedListLength takes precedence over searchBarModel", function () {
-      terria.searchBarModel.setTrait(
-        CommonStrata.definition,
-        "recommendedListLength",
-        10
-      );
+      runInAction(() => {
+        terria.configParameters.mergeValue(
+          ConfigStrata.user,
+          "searchBarConfig",
+          {
+            recommendedListLength: 10
+          }
+        );
+      });
       bingMapsSearchProvider.setTrait(
         CommonStrata.definition,
         "recommendedListLength",
@@ -65,11 +79,15 @@ describe("LocationSearchProvider", function () {
     });
 
     it(" - provider flightDurationSeconds takes precedence over searchBarModel", function () {
-      terria.searchBarModel.setTrait(
-        CommonStrata.definition,
-        "flightDurationSeconds",
-        3.0
-      );
+      runInAction(() => {
+        terria.configParameters.mergeValue(
+          ConfigStrata.user,
+          "searchBarConfig",
+          {
+            flightDurationSeconds: 3.0
+          }
+        );
+      });
       bingMapsSearchProvider.setTrait(
         CommonStrata.definition,
         "flightDurationSeconds",